# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Jan TodoList API",
    "description": "API specification for the Jan TodoList application, a simple authentication-based TODO list manager.",
    "version": "1.0.0",
    "contact": {
      "name": "Development Team"
    }
  },
  "servers": [
    {
      "url": "https://jan-todolist-api.railway.app/api",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development Server"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "API endpoints for user authentication (signup and login)."
    },
    {
      "name": "Todos",
      "description": "API endpoints for managing todo items."
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "User Signup",
        "description": "Registers a new user account.",
        "operationId": "signupUser",
        "requestBody": {
          "description": "User details for registration.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input or user already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "DuplicateEmail": {
                    "value": { "error": "이미 사용 중인 이메일입니다" }
                  },
                  "InvalidPassword": {
                    "value": { "error": "비밀번호는 8자 이상이어야 합니다" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "User Login",
        "description": "Logs in an existing user and returns a JWT.",
        "operationId": "loginUser",
        "requestBody": {
          "description": "User credentials for login.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "이메일 또는 비밀번호가 일치하지 않습니다"
                }
              }
            }
          }
        }
      }
    },
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get all todos for the current user",
        "description": "Retrieves a list of all todo items for the authenticated user, sorted by due date.",
        "operationId": "getTodos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "A list of todos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT is missing or invalid."
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Create a new todo",
        "description": "Adds a new todo item to the user's list.",
        "operationId": "createTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "description": "The todo item to create.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTodo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT is missing or invalid."
          }
        }
      }
    },
    "/todos/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "The ID of the todo item.",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "put": {
        "tags": ["Todos"],
        "summary": "Update an existing todo",
        "description": "Updates the content, completion status, or due date of a specific todo item.",
        "operationId": "updateTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "description": "The fields to update on the todo item.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden - User does not own this todo."
          },
          "404": {
            "description": "Not Found - Todo with this ID does not exist."
          }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Delete a todo",
        "description": "Permanently deletes a specific todo item.",
        "operationId": "deleteTodo",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": {
            "description": "Todo deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden - User does not own this todo."
          },
          "404": {
            "description": "Not Found - Todo with this ID does not exist."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "test@example.com"
          },
          "name": {
            "type": "string",
            "example": "홍길동"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-26T10:00:00Z"
          }
        }
      },
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "example": "프로젝트 보고서 작성"
          },
          "is_completed": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-11-27"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-26T10:30:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-11-26T10:30:00Z"
          }
        }
      },
      "NewUser": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "password123"
          },
          "name": {
            "type": "string",
            "example": "홍길동"
          }
        }
      },
      "Login": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          }
        }
      },
      "NewTodo": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "새로운 할일"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-11-30"
          }
        }
      },
      "UpdateTodo": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "수정된 할일 제목"
          },
          "is_completed": {
            "type": "boolean",
            "example": true
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2025-12-01"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MTc2ODc..."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}